@@ -1,41 +1,25 @@
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-import decoration from '../../resources/img/vision.png';
-import { useState } from "react";
-import ComicsList from "../comicsList/ComicsList";
-import AppBanner from "../appBanner/AppBanner";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { MainPage, ComicsPage } from "../pages";
 
 const App = () => {
-
-    const [selectedChar, setChar] = useState(null);
-
-    const onCharSelected = (id) => {
-        setChar(id);
-    }
-
     return (
-        <div className="app" >
-            <AppHeader />
-            <main>
-                {/* <ErrorBoundary>
-                    <RandomChar />
-                </ErrorBoundary>
-                <div className="char__content">
-                    <ErrorBoundary>
-                        <CharList onCharSelected={onCharSelected} />
-                    </ErrorBoundary>
-                    <ErrorBoundary>
-                        <CharInfo charId={selectedChar} />
-                    </ErrorBoundary>
-                </div>
-                <img className="bg-decoration" src={decoration} alt="vision" /> */}
-                <AppBanner />
-                <ComicsList />
-            </main>
-        </div>
+        <Router>
+            <div className="app" >
+                <AppHeader />
+                <main>
+                    <Switch>
+                        <Route exact path="/MarvelApp">
+                            <MainPage />
+                        </Route>
+                        <Route exact path='/MarvelApp/comics'>
+                            <ComicsPage />
+                        </Route>
+
+                    </Switch>
+                </main>
+            </div>
+        </Router>
     )
 }
 
