@@ -1,18 +1,19 @@
 import './appHeader.scss';
+import { Link, NavLink } from 'react-router-dom';
 
 const AppHeader = () => {
     return (
         <header className="app__header">
             <h1 className="app__title">
-                <a href="#">
+                <Link to="/MarvelApp">
                     <span>Marvel</span> information portal
-                </a>
+                </Link>
             </h1>
             <nav className="app__menu">
                 <ul>
-                    <li><a href="#">Characters</a></li>
+                    <li><NavLink exact activeStyle={{ 'color': '#9f0013' }} to="/MarvelApp">Characters</NavLink></li>
                     /
-                    <li><a href="#">Comics</a></li>
+                    <li><NavLink exact activeStyle={{ 'color': '#9f0013' }} to="/MarvelApp/comics">Comics</NavLink></li>
                 </ul>
             </nav>
         </header>
